Sicherheits- und Qualitaetsfixes aus dem Code-Review
Sicherheit:
* api.create_week_shipments: auf 10 Tage gedeckelt (MAX_WEEK_SHIPMENTS) und auf
  DPD-Sendungen beschraenkt. Der Endpunkt ist whitelisted und damit auch ohne
  den Button aufrufbar - jeder Tag ist ein echter, kostenpflichtiger Auftrag.
* api.create_shipment_from_delivery_note: sperrt den Lieferschein, bevor es
  nach einer vorhandenen Versandsendung sucht. Zwei parallele Aufrufe fanden
  sonst beide kein "existing", legten zwei Sendungen an und kauften zwei
  Etiketten; der for_update-Schutz in create_label() greift dort nicht, weil
  es zwei verschiedene Dokumente sind.
* check_permission("write") in allen Settings-Aktionen (Verbindungstest,
  Katalog aktualisieren, Abholorte laden, Saldo abgleichen). Sie loesen echte
  API-Aufrufe mit den hinterlegten Zugangsdaten aus bzw. legen Stammdaten an -
  das Leserecht, das eine whitelisted Dokumentmethode mitbringt, reicht nicht.
* DHL: Base64-Etiketten werden aus dem api_response entfernt (strip_label_data),
  sie haengen bereits als File am Dokument. api_request/api_response jetzt
  einheitlich bei 100.000 Zeichen gekappt, Kuerzung sichtbar markiert.
* DPD-Tracking: Sendungsnummer wird URL-kodiert, bevor sie in den Pfad geht.
* versandsendung.js: shipment_number vor dem Rendern escapen (kommt roh vom
  Carrier), analog zum bereits vorhandenen Escaping in delivery_note.js.
* absender.resolve(): ein deaktivierter Versandabsender wird auch dann
  abgelehnt, wenn er direkt an der Sendung haengt (z. B. vom Kunden geerbt).

Korrektheit:
* Versandsendung.on_cancel: Status und api_response per db_set schreiben.
  on_cancel laeuft in run_post_save_methods(), also nach dem db_update() -
  eine einfache Zuweisung landete nie in der DB, die Sendung blieb dauerhaft
  auf "Etikett erstellt" stehen.
* DHL-OAuth-Token-Cache haengt jetzt an Umgebung UND Zugangsdaten (gehashter
  Schluessel) statt an einem globalen Key. Ein Wechsel Sandbox<->Produktion
  oder eine Key-Rotation wirkt damit sofort statt erst nach Ablauf des alten
  Tokens. Dazu ein einmaliger Retry nach HTTP 401.
* DPD: WSDL-Cache (InMemoryCache) + Wiederverwendung des zeep-Clients. Vorher
  wurde das WSDL bei jedem Login und jedem storeOrders neu geladen.
* poll_open_shipments: Carrier-Liste kommt aus der Registry
  (carriers_with_tracking) statt hartkodiert, Sortierung nach aeltestem
  Update - sonst kam bei mehr offenen Sendungen als dem Limit immer dieselbe
  Teilmenge dran.
* _notify_problem: eine User-Abfrage statt eines get_value je Rolleninhaber.
* search_index auf Versandsendung.delivery_note.

Tests und CI:
* test_versandsendung.py von 4 auf 28 Tests: DHL-Payload-Bau (Mehrcolli,
  unvollstaendiger Absender, Retoure ohne Abrechnungsnummer), Gewichtslogik,
  Tracking-Statuslogik inkl. "Unbekannt ueberschreibt keinen bekannten
  Status", DPD-Abholtage. Ein Test gleicht die tracking_status-Optionen im
  DocType gegen die Konstanten in carriers/base.py ab.
* .github/workflows/ci.yml: ruff, Syntaxpruefung, DocType-JSON-Validierung
  inkl. Abgleich field_order <-> fields.
* pyproject.toml: BLE in die ruff-Auswahl aufgenommen, damit die vorhandenen
  "noqa: BLE001"-Kommentare nicht als unnoetig (RUF100) gemeldet werden.
  Tote "noqa: PLC0415"-Kommentare entfernt, restliche Funde gefixt - ruff
  laeuft jetzt sauber durch.

+++ versand_integration/carriers/dhl/mapper.py
@@ -182,9 +182,9 @@ def _details(weight_kg, length_cm, width_cm, height_cm):
 	if length_cm and width_cm and height_cm:
 		details["dim"] = {
 			"uom": "cm",
-			"length": int(round(flt(length_cm))),
-			"width": int(round(flt(width_cm))),
-			"height": int(round(flt(height_cm))),
+			"length": round(flt(length_cm)),
+			"width": round(flt(width_cm)),
+			"height": round(flt(height_cm)),
 		}
 	return details
